Converted ASCII passwords to use 7 bits per character instead of 8
Successfully tested with the default seed length in-game. Also needed to change the ascii seed tests, since converting characters to use 7 bits instead of 8 caused changes to resulting hex values when converting between ascii and hex. This should not effect the user, only the tests.
Additionally, this increases the max character length from 16 to 18.